# Add a second CUDA transform reduction strategy for small reduction sizes