--- a/llvm-external-projects/iree-dialects/include/iree-dialects/Transforms/TransformMatchers.h
+++ b/llvm-external-projects/iree-dialects/include/iree-dialects/Transforms/TransformMatchers.h
@@ -570,6 +570,8 @@ private:
 
 struct MatchedReductionCaptures {
   int64_t reductionRank = 0;
+  // TODO: capture all dims.
+  int64_t mostMinorParallelDimensionSize = 0;
   int64_t reductionDimensionSize = 0;
   int64_t maybeLeadingRank = 0;
   int64_t maybeTrailingRank = 0;