This revision adds a new strategy for the small reduction domain.
For small reductions, we rely on L1 cache locality and do not try to parallelize individual reductions: every thread performs the full reduction locally.

This is empirically determined profitable up until a reduction size of 64.
This can be fine-tuned later when we have better support for filtering and profitability considerations.

Running the iree-samples/transform_dialect script:
```
(  benchmark-transform-create -r  ../iree-samples/transform_dialect/benchmark_linalg_reductions.stub.mlir   reduction_2d_static 1234567 15 )
```

before:
```
With transform dialect: reduction_2d_static --function_input="1234567x15xf32" P50: 1278300.0000 ns 14.48682234217319877962 GElements/s
Without transform dialect: reduction_2d_static --function_input="1234567x15xf32" P50: 2778900.0000 ns 6.66396955629925510093 GElements/s
```

with this PR:
```
With transform dialect: reduction_2d_static --function_input="1234567x15xf32" P50: 150060.00 ns 123.40733706517393042782 GElements/s
Without transform dialect: reduction_2d_static --function_input="1234567x15xf32" P50: 2808400.0000 ns 6.59396987608602763139 GElements/s
```